requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
clr==1.0.3
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.1
decorator==5.1.1
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
Markdown==3.5.2
MarkupSafe==2.1.5
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.58.1
numpy==1.25.2
packaging==23.2
pandas==2.2.1
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==10.2.0
pip==22.0.2
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywaffle==1.1.0
PyWavelets==1.5.0
pyzmq==25.1.2
requests==2.31.0
scipy==1.11.4
seaborn==0.12.2
setuptools==59.6.0
six==1.16.0
sqlparse==0.4.4
tangled-up-in-unicode==0.2.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
widgetsnbextension==4.0.10
wordcloud==1.9.3
xgboost==2.0.3